chore(refactor): accordion folder structure
