+++ src/components/ui/TripCardAccordion/AccordionDetails/TripCardDetails.styled.ts
