+++ src/components/ui/TripCardAccordion/AccordionDetails/TripCardDetails.tsx
