+++ src/components/ui/TripCardAccordion/AccordionDetails/components/Checkout.tsx
