+++ src/components/ui/TripCardAccordion/AccordionDetails/components/DetailItem/DetailItem.tsx
