+++ src/components/ui/TripCardAccordion/AccordionDetails/components/DetailsGrid.tsx
