+++ src/components/ui/TripCardAccordion/AccordionDetails/components/SeatLegend/SeatLegend.styled.ts
