+++ src/components/ui/TripCardAccordion/AccordionDetails/components/SeatLegend/SeatLegend.tsx
